rename packages; rebase to 6.0.0-snapshot and Spark 2.1.3; clean up pom.xml
