+++ .travis.yml
@@ -1,5 +1,5 @@
 #
-# Copyright © 2016-2017 Cask Data, Inc.
+# Copyright © 2016-2019 Cask Data, Inc.
 # 
 # Licensed under the Apache License, Version 2.0 (the "License"); you may not
 # use this file except in compliance with the License. You may obtain a copy of
@@ -16,11 +16,10 @@
 language: java
 
 jdk:
-  - oraclejdk7
   - oraclejdk8
 
 env:
-  - MAVEN_OPTS="-Xmx4096m -XX:MaxPermSize=128m" HADOOP_VERSION="2.4.1"
+  - MAVEN_OPTS="-Xmx4096m -XX:MaxPermSize=128m"
 
 sudo: false
