--- market/1.0.0/spec.json
@@ -1,40 +1,0 @@
-{
-  "specVersion": "1.0",
-  "label": "RDD Repartitioner",
-  "description": "Repartitions a RDD.",
-  "author": "Cask",
-  "org": "Cask Data, Inc.",
-  "created": 1488005294,
-  "categories": [ 
-    "hydrator-plugin"
-  ],
-  "cdapVersion": "[4.1.0,4.2.0-SNAPSHOT)",
-  "actions": [
-    {
-      "type": "one_step_deploy_plugin",
-      "label": "Deploy Repartitioner",
-      "arguments": [
-        {
-          "name": "name",
-          "value": "repartitioner"
-        },
-        {
-          "name": "version",
-          "value": "1.0.0"
-        },
-        {
-          "name": "scope",
-          "value": "user"
-        },
-        {
-          "name": "config",
-          "value": "repartitioner-1.0.0.json"
-        },
-        {
-          "name": "jar",
-          "value": "repartitioner-1.0.0.jar"
-        }
-      ]
-    }
-  ]
-}
