+++ src/test/java/io/cdap/plugin/repartitioner/RepartitionerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2017 Cask Data, Inc.
+ * Copyright © 2017-2019 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package org.example;
+package io.cdap.plugin.repartitioner;
 
 /**
  * Test for Repartitioner plugin.
